Refactor to separate animation and rendering
Rendering the pipes to the screen and animating a frame are two separate
concepts. Rendering takes care of drawing the pipes and flushing
everything to the terminal. Animating involves rendering a frame,
determining the delay required before the next frame needs rendering and
sleeping for that length of time.

This commit separtes the two concerns out. The `animate` function now
calls an arbitrary renderer, which is passed as a function pointer. The
`animate` function could call the rendering function directly, but in
the future we may want to pass a list of renders to allow for overlays.
